% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_df_from_gv_xlsx.R
\name{extract_df_from_gv_xlsx}
\alias{extract_df_from_gv_xlsx}
\title{Extract data.frame from Genvestigator .xlsx file}
\usage{
extract_df_from_gv_xlsx(file_path)
}
\arguments{
\item{file_path}{.xlsx file path}
}
\description{
Extract data.frame from Genvestigator .xlsx file
}
