% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_na_index.R
\name{extract_na_col_idx}
\alias{extract_na_col_idx}
\title{extract column index}
\usage{
extract_na_col_idx(df)
}
\arguments{
\item{df}{data.frame from read_gv_xlsx()}
}
\description{
extract column index
}
