% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gv2df.R
\name{gv2info_df}
\alias{gv2info_df}
\title{extract meta informations of experiments}
\usage{
gv2info_df(df)
}
\arguments{
\item{df}{data.frame from read_gv_xlsx()}
}
\description{
extract meta informations of experiments
}
