% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gv2df.R
\name{gv2meta_df}
\alias{gv2meta_df}
\title{extract meta informations of genes (e.g. Gene Symbol, Mearsure (probe ID))}
\usage{
gv2meta_df(df)
}
\arguments{
\item{df}{data.frame from read_gv_xlsx()}
}
\description{
extract meta informations of genes (e.g. Gene Symbol, Mearsure (probe ID))
}
