% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gv2df.R
\name{gv2tidy_data_df}
\alias{gv2tidy_data_df}
\title{extract data informations of experiments}
\usage{
gv2tidy_data_df(df)
}
\arguments{
\item{df}{data.frame from read_gv_xlsx()}
}
\description{
extract data informations of experiments
}
