% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_def.R
\name{param_def}
\alias{param_def}
\title{parameter difinition}
\usage{
param_def(df, meta_df)
}
\arguments{
\item{df}{data.frame from read_gv_xlsx()}

\item{meta_df}{data.frame from gv2meta_df()}
}
\description{
parameter difinition
}
