% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gv_xlsx.R
\name{read_gv_xlsx}
\alias{read_gv_xlsx}
\title{read xlsx exported from GeneVestigator}
\usage{
read_gv_xlsx(fpath)
}
\arguments{
\item{fpath}{file path}
}
\description{
read xlsx exported from GeneVestigator
}
